Refactor tests out to tests directory
