+++ pytest.ini
@@ -1,8 +1,6 @@
 [pytest]
 minversion = 8.4
-testpaths = src
-python_files = test_*.py *_test.py
-python_classes = Test
-python_functions = test_*
-addopts = -p no:warnings
-pythonpath = src
+testpaths = tests
+
+# -v for verbose, -s to show print statements
+addopts = -v
