+++ tests/document_chunker/data/textract_response_template.json
