+++ tests/document_chunker/test_bbox_utils.py
@@ -4,7 +4,7 @@ from typing import List, cast
 import pytest
 from textractor.entities.bbox import BoundingBox
 
-from document_chunker.utils.bbox_utils import combine_bounding_boxes
+from src.document_chunker.utils.bbox_utils import combine_bounding_boxes
 
 
 @dataclass
